SQL Trigger for Vendor Update Product Disassociation
When a vendor is DELETEd the Products remain available to the Customers to Order. This trigger handled that in basic form but can be customized to suit specific rules./ Similarly I have create a Trigger for Status Updates. I have handle it on the DB because of the size of the DATABASES we run and the efficiency of handling it rather on the data tier. This can be converted to functions and usedin the Application Tier.